pyproject.toml
[tool.poetry]
name = "dw_blog"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
alembic = "^1.12.1"
asyncpg = "^0.28.0"
sqlmodel = "^0.0.11"
uvicorn = "^0.23.2"
httpx = "^0.25.1"
pytest = "^7.4.3"
sqlalchemy-utils = "^0.41.1"
psycopg2-binary = "^2.9.9"
pytest-asyncio = "^0.21.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["pycryptodome"], version = "^3.3.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
